rename cli.py to _cli.py to imply private
